fixed incoming changes for stacks and specs to work correctly, moving events and statuses to separate classes, added spec for actual errors in events, updated vcrs
